backend/requirements.txt
claude-agent-sdk
fastapi
uvicorn[standard]
python-dotenv
pydantic
